+++ lib/angular2-tree-component.ts
@@ -6,6 +6,7 @@ import { ITreeOptions } from './defs/api';
 import { KEYS } from './constants/keys';
 import { TreeModel } from './models/tree.model';
 import { TreeNode } from './models/tree-node.model';
+import { TreeDraggedElement } from './models/tree-dragged-element.model';
 import { LoadingComponent } from './components/loading.component';
 import { LoadingComponent as DeprecatedLoadingComponent } from './components/deprecated-loading.component';
 import { TreeComponent } from './components/tree.component';
@@ -13,6 +14,8 @@ import { TreeNodeComponent } from './components/tree-node.component';
 import { TreeNodeContent } from './components/tree-node-content.component';
 import { TreeNodeContent as DeprecatedTreeNodeContent } from './components/deprecated-tree-node-content.component';
 import { TreeNodeDropSlot } from './components/tree-node-drop-slot.component';
+import { TreeDropDirective } from './directives/tree-drop.directive';
+import { TreeDragDirective } from './directives/tree-drag.directive';
 import { AdHocComponentFactoryCreator } from './components/adhoc-component-factory.service';
 
 import './polyfills';
@@ -21,6 +24,7 @@ import { deprecated } from './deprecated';
 export {
   TreeModel,
   TreeNode,
+  TreeDraggedElement,
   ITreeOptions,
   TREE_ACTIONS,
   KEYS,
@@ -30,6 +34,8 @@ export {
   TreeComponent,
   TreeNodeComponent,
   TreeNodeContent,
+  TreeDropDirective,
+  TreeDragDirective,
   TreeNodeDropSlot
 };
 
@@ -40,29 +46,38 @@ export {
     TreeNodeComponent,
     TreeNodeDropSlot,
     TreeNodeContent,
+    TreeDropDirective,
+    TreeDragDirective
   ],
   exports: [
     TreeComponent,
+    TreeDropDirective,
+    TreeDragDirective
   ],
   imports: [
     CommonModule,
   ],
+  providers: [
+    TreeDraggedElement
+  ]
 })
 export class TreeModule {}
 @NgModule({
   declarations: [
     DeprecatedLoadingComponent,
-    DeprecatedTreeNodeContent,
+    DeprecatedTreeNodeContent
   ],
   exports: [
     TreeComponent,
+    TreeDropDirective,
+    TreeDragDirective
   ],
   imports: [
     CommonModule,
     TreeModule,
   ],
   providers: [
-    AdHocComponentFactoryCreator,
+    AdHocComponentFactoryCreator
   ],
 })
 export class DeprecatedTreeModule {